Added lifetime annotation for ArchiveEntryIterator.
Previously the ArchiveEntryIterator returned by the
TarArchiveRef::entries() function was linked to the lifetime of the
TarArchiveRef by means of the self reference. The ArchiveEntryIterator
however only depends on the underlying data not the reference. Thus it
should have the lifetime 'a.

A problem occures when your data lives longer than the TarArchiveRef,
for example if the data is 'static but the archive reference only on the
stack. Now the data is valid for 'static but the data returned by the
ArchiveEntryIterator only for the liftime of the stack variable.